feat: adds incomplete shadow map
Note: Not only is the vs not working yet, this is supposed to be a shadow map for every light, but thats impossible! So once its working, I will turn this into numLights number of shadow maps. For now it loads and runs in Vulkan without breaking the other 2 passes though.